+++ Icarus Protocol/Assets/UI/Manual/ManualUIContainer .prefab
@@ -295,7 +295,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &4097881539480462667
 RectTransform:
   m_ObjectHideFlags: 0
@@ -366,11 +366,11 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 1790094505410326034, guid: 66d53311be30bd2418448a97fd3041fb, type: 3}
       propertyPath: m_fontSize
-      value: 16
+      value: 20
       objectReference: {fileID: 0}
     - target: {fileID: 1790094505410326034, guid: 66d53311be30bd2418448a97fd3041fb, type: 3}
       propertyPath: m_fontSizeBase
-      value: 16
+      value: 20
       objectReference: {fileID: 0}
     - target: {fileID: 6697249966062979928, guid: 66d53311be30bd2418448a97fd3041fb, type: 3}
       propertyPath: m_Pivot.x
