+++ Icarus Protocol/Assets/UI/StartMenu/StartButtonBehavior.cs
@@ -13,21 +13,10 @@ using Image = UnityEngine.UI.Image;
 /// </summary>
 public class StartButtonBehavior : MonoBehaviour
 {
-
-    private TMP_Text text;
-    private FontWeight startingWeight;
-
     /// <summary>
     /// A reference to the UI Handler for facilitating layout transitions.
     /// </summary>
     public UILayoutHandler UiHandler;
-
-    private void Start()
-    {
-        text = GetComponentInChildren<TMP_Text>();
-        startingWeight = text.fontWeight;
-
-    }
 
     /// <summary>
     /// Called when the button is clicked.
@@ -35,17 +24,5 @@ public class StartButtonBehavior : MonoBehaviour
     public void OnClick()
     {
         UiHandler.TransitionToLevelSelect(0);
-    }
-
-    public void OnHover()
-    {
-        text.color = Color.black;
-        text.fontWeight = FontWeight.Bold;
-    }
-
-    public void OffHover() 
-    {
-        text.color = Color.white;
-        text.fontWeight = startingWeight;
     }
 }
