+++ Icarus Protocol/Assets/UI/UICommon/HoverBehavior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c0ee8df0620a25142a3520af60fdbe61
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
